requirements.txt
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
instaloader==4.15
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
soupsieve==2.8.3
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
Werkzeug==3.1.5
